requirements.txt
fastapi==0.112.2
uvicorn[standard]==0.30.6
httpx==0.27.2
beautifulsoup4==4.12.3
selectolax==0.3.21
readability-lxml==0.8.1
lxml==5.3.0
html5lib==1.1
markdownify==0.12.1
python-dotenv==1.0.1
sentence-transformers==3.0.1
faiss-cpu==1.8.0.post1
typesense==0.21.0
rapidfuzz==3.9.6
numpy==1.26.4
pydantic==2.8.2
